Adde Save FIle Locally and to the Db

+++ WebAPI/EFTest/EFTest/Data/FileHandlerService.cs
@@ -221,14 +221,14 @@ public class FileHandlerService
         var filePath = Path.Combine(_basePath, mFile.FileAddress);
 
         //chk if file is local directory
-        //if (!File.Exists(filePath))
-        //{
-        //    throw new ArgumentException("File does not exists in local directory");
-        //}
-        //else
-        //{
+        if (!File.Exists(filePath))
+        {
+            throw new ArgumentException("File does not exists in local directory");
+        }
+        else
+        {
             File.Delete(filePath);
-        //}
+        }
 
         //remove from db
         _appDbContext.MyFiles.Remove(mFile);
